% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_prob.R
\name{dir_prob}
\alias{dir_prob}
\title{Compute Dirchlet Probability of Last Item in a Sequence}
\usage{
dir_prob(x)
}
\arguments{
\item{x}{The sequence passed as a Factor}

\item{alpha}{smoothing factor applied to counts}
}
\value{
Probability of seeing the last item in the sequence given the preceding elements
}
\description{
The probability of observing the last (jth) element in a sequence.

\eqn{p(x^{j} = k | X^{j-1}, \alpha) = \frac{n_{k}^{j-1}+1}{j-1+K}}
}
\examples{
dir_prob(factor(x, levels = 1:3))
}
