% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_prob.R
\name{dir_prob1}
\alias{dir_prob1}
\title{Compute Dirchlet Probability of Last Item in a Sequence}
\usage{
dir_prob1(x, n = 1, k = -1)
}
\arguments{
\item{x}{The sequence}

\item{n}{Number of element from the end of the sequence to consider (0 or 1)}

\item{k}{Which element to calculate the probability for}
}
\value{
Probability of seeing item \code{k} in the sequence given the preceding elements up to \code{n} elements from the end
}
\description{
A general case of@seealso \code{\link[=dir_prob]{dir_prob()}}
}
\examples{
dir_prob(factor(c(1), levels = 1:3)) == dir_prob1(factor(c(1), levels = 1:3))
}
