% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictability.R
\name{predictability}
\alias{predictability}
\title{Calculate predictability in terms of negative entropy}
\usage{
predictability(x)
}
\arguments{
\item{x}{The sequence of type Factor with specified levels}
}
\value{
Predictability score
}
\description{
Predictability is quantified it terms of negative entropy

\eqn{-H(p^{j}) = \sum_{k=1}^{K}p(x^j = k|X^j, \alpha)\log_{2}p(x^j = k|X^j, \alpha)}

Note that, different from surprise, here, predictability is estimated considering also the event j, and not just up to j − 1

\eqn{p(x^j=k)} values are calculated in \code{\link[=prob_target]{prob_target()}}
}
\examples{
predictability(factor(c(1), levels = 1:3))
}
