% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surprise.R
\name{surprise}
\alias{surprise}
\title{Compute the surprise of observing the last event in a sequence}
\usage{
surprise(x)
}
\arguments{
\item{x}{}
}
\value{
Information in bits
}
\description{
Following the formula for Shannon information:

\eqn{I(x^{j} = k) = -log_{2}p(x^{j} = k | X^{j-1}, \alpha)}

Note that the probability \eqn{p(x^{j} = k | X^{j-1}, \alpha)} is the definition
for Dirichlet probability seen in
}
\examples{
surprise(factor(x, levels = 1:3))
}
\seealso{
\code{\link[=dir_prob]{dir_prob()}}
}
